* commit '990e529a05abe8b511fccf79aaf7861633484d1f':
  add endpoint for getting a device by name
  add some docstrings and reformat
  try to be more consistent with docstring format
  get test_db.py working with the Device model
  fix more pylint errors
  fix some pylint errors
  don't use update_from_heartbeat in test_device.py update_from_heartbeat gets tested indirectly in test_heartbeat.py
  fix typo in parameter name
  add test for device/<device_id> endpoint with bad ID
  add endpoint for getting device by id, allow filtering device list by state add tests for device endpoints

+++ src/app/controller/schemas/device.py
@@ -47,7 +47,7 @@ CAMERA_STATUS = Model('camera_status', {
             "the recording session in seconds"
         )
     ),
-    'fps': fields.Integer(
+    'fps': fields.Float(
         description=(
             "If the camera is recording, this value indcates how many frames "
             "per second are being captured"
